DT-861 Refactor on scanner pipeline
